backend/requirements.txt
fastapi
uvicorn[standard]
neo4j
python-dotenv
pydantic
pydantic-settings
anthropic
sentence-transformers
numpy
scikit-learn
chromadb
feedparser
requests
lxml 